perculator_agent_web/requirements.txt
fastapi
uvicorn
openai
tree_sitter
tree_sitter_rust
faiss-cpu
sentence-transformers
jinja2